rename(focus-trap): FocusTrapModel filename
