+++ packages/common/focus-trap/lib/__tests__/initialization.test.ts
@@ -1,5 +1,5 @@
 import {screen} from '@testing-library/dom';
-import type {FocusTrapModel} from '../main';
+import type {FocusTrapModel} from '../FocusTrapModel';
 import {observableFocusTrap} from './observableFocusTrap';
 
 let trap: FocusTrapModel | undefined;
